Refactor interface UO (#19229)
